Crate bp_asset_hub_rococo

Module with configuration which reflects AssetHubRococo runtime setup.

Structs

AccountInfoStorageMapKeyProvider

Provides a storage key for account data.

AssetHubRococo

AssetHubRococo parachain.

BlockExecutionWeight

Importing a block with 0 Extrinsics.

BlockLength

Size limit of the Cumulus-based bridge hub blocks.

BlockWeights

Weight limit of the Cumulus-based bridge hub blocks.

BlockWeightsForAsyncBacking

Weight limit of the Cumulus-based bridge hub blocks when async backing is enabled.

CreateForeignAssetDeposit

Should match the AssetDeposit of the ForeignAssets pallet on Asset Hub.

ExtrinsicBaseWeight

Executing a NO-OP System::remarks Extrinsic.

Perbill

A fixed point representation of a number in the range [0, 1].

XcmBridgeHubRouterTransactCallMaxWeight

Some sane weight to execute xcm::Transact(pallet-xcm-bridge-hub-router::Call::report_bridge_status).

Enums

Call

AssetHubRococo Runtime Call enum.

XcmBridgeHubRouterCall

A minimized version of pallet-xcm-bridge-hub-router::Call that can be used without a runtime.

Constants

ASSET_HUB_ROCOCO_PARACHAIN_ID

Identifier of AssetHubRococo in the Rococo relay chain.

AVERAGE_BLOCK_INTERVAL

Average block time for Cumulus-based parachains

AVERAGE_ON_INITIALIZE_RATIO

We assume that ~5% of the block weight is consumed by on_initialize handlers. This is used to limit the maximal weight of a single extrinsic.

BEST_FINALIZED_ASSET_HUB_ROCOCO_HEADER_METHOD

Name of the <ThisChain>FinalityApi::best_finalized runtime method.

EXTRA_STORAGE_PROOF_SIZE

Number of extra bytes (excluding size of storage value itself) of storage proof, built at Polkadot-like chain. This mostly depends on number of entries in the storage trie. Some reserve is reserved to account future chain growth.

FREE_HEADERS_INTERVAL_FOR_ASSET_HUB_ROCOCO_METHOD

Name of the <ThisChain>FinalityApi::free_headers_interval runtime method.

FROM_ASSET_HUB_ROCOCO_MESSAGE_DETAILS_METHOD

Name of the From<ThisChain>InboundLaneApi::message_details runtime method.

MAXIMUM_BLOCK_WEIGHT

We allow for 0.5 seconds of compute with a 6 second average block time.

MAXIMUM_BLOCK_WEIGHT_FOR_ASYNC_BACKING

We allow for 2 seconds of compute with a 6 second average block.

MAX_ASSET_HUB_HEADER_SIZE

Maximal asset hub header size.

MAX_BRIDGE_HUB_HEADER_SIZE

Maximal bridge hub header size.

MAX_UNCONFIRMED_MESSAGES_IN_CONFIRMATION_TX

Maximal number of unconfirmed messages at inbound lane for Cumulus-based parachains. Note: this value is security-relevant, decreasing it should not be done without careful analysis (like the one above).

MAX_UNREWARDED_RELAYERS_IN_CONFIRMATION_TX

Maximal number of unrewarded relayer entries at inbound lane for Cumulus-based parachains. Note: this value is security-relevant, decreasing it should not be done without careful analysis (like the one above).

NORMAL_DISPATCH_RATIO

We allow Normal extrinsics to fill up the block up to 75%, the rest can be used by Operational extrinsics.

SLOT_DURATION

TO_ASSET_HUB_ROCOCO_MESSAGE_DETAILS_METHOD

Name of the To<ThisChain>OutboundLaneApi::message_details runtime method.

TX_EXTRA_BYTES

Maximal number of bytes, included in the signed Polkadot-like transaction apart from the encoded call itself.

WITH_ASSET_HUB_ROCOCO_MESSAGES_PALLET_NAME

Name of the With-AssetHubRococo messages pallet instance that is deployed at bridged chains.

WITH_ASSET_HUB_ROCOCO_RELAYERS_PALLET_NAME

Name of the With-AssetHubRococo bridge-relayers pallet instance that is deployed at bridged chains.

WITH_BRIDGE_ROCOCO_TO_WESTEND_MESSAGES_PALLET_INDEX

Pallet index of BridgeWestendMessages: pallet_bridge_messages::<Instance1>.

Traits

AssetHubRococoFinalityApi

API for querying information about the finalized chain headers.

FromAssetHubRococoInboundLaneApi

Inbound message lane API for messages sent by this chain.

ToAssetHubRococoOutboundLaneApi

Outbound message lane API for messages that are sent to this chain.

Functions

build_congestion_message

Builds an (un)congestion XCM program with the report_bridge_status call for ToWestendXcmRouter.

Type Aliases

AccountId

Id of account on Polkadot-like chains.

AccountPublic

Public key of account on Polkadot-like chains.

AccountSigner

Public key of the chain account that may be used to verify signatures.

Address

The address format for describing accounts.

Balance

The balance of an account on Polkadot-like chain.

BlockNumber

Block number type used in Polkadot-like chains.

Hash

Hash type used in Polkadot-like chains.

Hasher

The type of object that can produce hashes on Polkadot-like chains.

Hashing

Hashing type.

Header

The header type used by Polkadot-like chains.

Nonce

Nonce of a transaction on the Polkadot-like chains.

Signature

Signature type used by Polkadot-like chains.

SignedBlock

Polkadot-like block signed with a Justification.

TransactionExtension

Signed extension that is used by all bridge hubs.

UncheckedExtrinsic

Unchecked Extrinsic type.